Moved Windows specific components to Windows folder
